🔧 Refactor QCO operations to improve static qubit handling: update StaticOp to use Pure memory effects, introduce SSABeforeForDeallocOrder for deterministic sorting of SSA values, and enhance conversion patterns for static qubits across various dialects. Add new utility functions for better qubit management in tests.

+++ mlir/lib/Conversion/QCOToQC/QCOToQC.cpp
@@ -156,8 +156,7 @@ struct ConvertQCOStaticOp final : OpConversionPattern<qco::StaticOp> {
   matchAndRewrite(qco::StaticOp op, OpAdaptor /*adaptor*/,
                   ConversionPatternRewriter& rewriter) const override {
     // Create qc.static with the same index
-    rewriter.replaceOpWithNewOp<qc::StaticOp>(
-        op, static_cast<int64_t>(op.getIndex()));
+    rewriter.replaceOpWithNewOp<qc::StaticOp>(op, op.getIndex());
     return success();
   }
 };
